add HealthCare Service Tables

+++ PADC-FP-HealthCareDirectory/app/src/main/java/com/padc/healthcaredirectory/data/vos/OperationVO.java
@@ -17,7 +17,7 @@ public class OperationVO {
         return operationId;
     }
 
-    public void setOperationId(int operationId) {
+    public void setOperationId(long operationId) {
         this.operationId = operationId;
     }
