+++ PADC-FP-HealthCareDirectory/app/src/main/java/com/padc/healthcaredirectory/data/vos/SpecialityVO.java
@@ -17,7 +17,7 @@ public class SpecialityVO {
         return specialityId;
     }
 
-    public void setSpecialityId(int specialityId) {
+    public void setSpecialityId(long specialityId) {
         this.specialityId = specialityId;
     }
